+++ code/ch14/optimized_sliding_window_demo.py
@@ -444,8 +444,8 @@ class SlidingWindowDemoBenchmark(VerificationPayloadMixin, BaseBenchmark):
         """Setup: Initialize SDPA-based attention module."""
         torch.manual_seed(42)
         
-        # Use bfloat16 for wide hardware support and stable Flash Attention
-        self.dtype = torch.bfloat16
+        # Match baseline dtype for strict signature/workload comparability.
+        self.dtype = torch.float16
         
         self.model = OptimizedSDPAAttention(
             embed_dim=self.embed_dim,
